compelted garage lecture notes

+++ Unit_five/moviesDB/controllers/movie.controller.js
@@ -1,17 +1,19 @@
 const router = require('express').Router();
 const Movie = require('../models/movie.models');
 const { errorResponse } = require('../utils'); //will target index by default if no other file is targetted
+const validateSession = require('../middleware/validate-session');
 
 /////toDo POST
-router.post('/', async(req,res) => {
+router.post('/', validateSession, async(req,res) => {
     try {
         //*1) pull data from client (body)
+        // console.log(req);
         const { 
             title, genre, rating, length, releaseYear
          } = req.body;
         //*2) create a new object using the model
          const movie = new Movie({
-            title, genre, rating, length, releaseYear
+            title, genre, rating, length, releaseYear,  owner_id: req.user._id
          }); //shaping the object the same as the pulled data simplifies the code if the object will be the same as schema/data pulled
         //*3) save the object to the DB
          const newMovie = await movie.save();
@@ -26,7 +28,7 @@ router.post('/', async(req,res) => {
     }
 });
 /////toDo GET ALL
-router.get('/', async(req,res) => {
+router.get('/', validateSession, async(req,res) => {
     try {
         // const allMovies = await Movie.find({});
         // if(!allMovies || allMovies.length === 0){
@@ -37,7 +39,11 @@ router.get('/', async(req,res) => {
         //         allMovies
         //     })
         // } this works as well
+        
+
         const allMovies = await Movie.find();
+
+        // console.log(req.user);
 
         allMovies ?
             res.status(200).json({
@@ -51,7 +57,7 @@ router.get('/', async(req,res) => {
     }
 });
 /////toDo Get One
-router.get('/find-one/:id', async(req,res) => {
+router.get('/find-one/:id', validateSession, async(req,res) => {
     try {
         // const movieId = await Movie.findOne({_id: _id});
         // if(!movieId) {
@@ -61,8 +67,16 @@ router.get('/find-one/:id', async(req,res) => {
         //         message: "found by id",
         //         movieId
         //     })not surer if this works
-        const { id } = req.params;
-        const getMovie = await Movie.findOne({_id: id});
+
+
+        // const { id } = req.params;
+        const filter = {
+            id: req.params,
+            _id: req.params.id,
+            owner_id: req.user._id
+        }
+
+        const getMovie = await Movie.find(filter);
 
         if(!getMovie) throw new Error('no movie found');
 
@@ -75,7 +89,7 @@ router.get('/find-one/:id', async(req,res) => {
     }
 });
 /////toDO GET ALL by Genre
-router.get('/genre/:genre', async(req,res) => {
+router.get('/genre/:genre', validateSession, async(req,res) => {
     try {
         const { genre } = req.params;
         let pedroPascal; //set required genre entered to be pascal case or first uppercase and rest lower case
@@ -96,7 +110,7 @@ router.get('/genre/:genre', async(req,res) => {
                 pedroPascal += genre[i].toLowerCase();
         }
         }
-        const movieGenre = await Movie.find({genre: pedroPascal});
+        const movieGenre = await Movie.find({genre: pedroPascal, owner_id: req.user._id});
 
         if (movieGenre.length === 0) throw new Error('No genre found');
 
@@ -111,16 +125,21 @@ router.get('/genre/:genre', async(req,res) => {
 //toDO PATCH/PUT One
 //patch is for smoll put is for big
 //patch addresses specific item -- put addresses the whole document
-router.patch('/:id', async(req,res) => {
+router.patch('/:id', validateSession, async(req,res) => {
     try {
         //1) pull value from parameter
-        const { id } = req.params;
+        // const { id } = req.params;
+        const filter = {
+            _id: req.params.id,
+            owner_id: req.user._id
+        }
         //2) pull data from the body
         const info = req.body;
         //3) use method to locate a document based off the ID and pass in new info
         //* .findOneAndUpdate(query, document, options)
         const returnOption = {new: true}; //option - returns the updated document
-        const updated = await Movie.findOneAndUpdate({_id: id}, info, returnOption);
+        // const updated = await Movie.findOneAndUpdate({_id: id}, info, returnOption);
+        const updated = await Movie.findOneAndUpdate(filter, info, returnOption);
         //4) respond to client
         res.status(200).json({
             results:updated
@@ -130,13 +149,13 @@ router.patch('/:id', async(req,res) => {
     }
 });
 //toDO DELETE One
-router.delete('/:id', async(req,res) => {
+router.delete('/:id', validateSession, async(req,res) => {
     try {
         //1) capture data (ID)
         const { id } = req.params;
         
         //2) use a delete method to locate and remove
-        const deleteMovie = await Movie.deleteOne({_id: id});
+        const deleteMovie = await Movie.deleteOne({_id: id, owner_id: req.user._id});
         console.log(deleteMovie);
         //3) respond to client
         deleteMovie.deletedCount ?
